Sidebranch

+++ models/product.js
@@ -1,6 +1,8 @@
 const mongodb = require('mongodb');
 const getDb = require('../util/database').getDb;
 
+const item_per_page = 2;
+
 class Product{
   constructor(title, price, description, imageUrl, id, userId){
     this.title = title;
@@ -29,9 +31,12 @@ class Product{
     });
   }
 
-  static fetchAll(){
+  static fetchAll(page){
     const db = getDb();
-    return db.collection('products').find().toArray()
+    return db.collection('products').find()
+    .skip((page - 1) * item_per_page)
+    .limit(item_per_page)
+    .toArray()
     .then(products => {
       // console.log(products);
       return products;
@@ -41,6 +46,14 @@ class Product{
     });
     //find doesn't return a promise it return a cursor,
     //an object provided by mongodb which allows us to go through our elements, our documents step by step.
+  }
+  static findProduct(id){
+    const db = getDb();
+    return db.collection('products').find({userId : id}).toArray()
+    .then(products => {
+      return products;
+    })
+    .catch(err => console.log(err));
   }
 
   static findById(prodId){
@@ -56,9 +69,9 @@ class Product{
       console.log(err);
     });
   }
-  static deleteById(prodId){
+  static deleteById(prodId, id){
     const db = getDb();
-    return db.collection('products').deleteOne({_id: new mongodb.ObjectId(prodId)})
+    return db.collection('products').deleteOne({_id: new mongodb.ObjectId(prodId), userId : id})
     .then(result => {
       console.log('Deleted');
     })
